snapshot of changes for OVPMS-1960 . make attachments viewable when editing the claim . suppress display of the attachment id node
svn path=/projects/openvpms/branches/OVPMS-1960/; revision=7713

+++ openvpms-web/openvpms-web-workspaces/src/main/java/org/openvpms/web/workspace/patient/insurance/claim/AttachmentCollectionEditor.java
@@ -147,7 +147,8 @@ class AttachmentCollectionEditor extends ActRelationshipCollectionEditor impleme
         if (!object.isNew()) {
             service.remove(object);
         }
-        if (!reference.isNew()) {
+        if (reference != null && !reference.isNew()) {
+            // TODO - would be improved by OBF-247
             Document document = (Document) service.get(reference);
             if (document != null) {
                 service.remove(document);
